+++ tests/testthat/test-get_field_index.R
@@ -6,12 +6,15 @@ test_that("gets df from connection", {
   expect_true(is.data.frame(get_field_index(conn)))
 })
 
-test_that("returns liesWithinLayer +  shortDescription when verbose", {
+test_that("returns shortDescription when verbose", {
   index <- get_field_index(conn, verbose = TRUE)
-  expect_true("liesWithinLayer" %in% colnames(index))
   expect_true("shortDescription" %in% colnames(index))
 })
 
+test_that("returns liesWithinLayer when find_layers", {
+  index <- get_field_index(conn, find_layers = TRUE)
+  expect_true("liesWithinLayer" %in% colnames(index))
+})
 
 test_that("returns Place when gather_trenches", {
   index <- get_field_index(conn, gather_trenches = TRUE)